csgo_discount_parser/requirements.txt
aiogram==3.0.0b7
aiohttp==3.8.4
python-dotenv==1.0.0
